src/embedder/requirements.txt
fastapi==0.109.2
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pgvector==0.3.6
https://download.pytorch.org/whl/cpu/torch-2.3.0%2Bcpu-cp311-cp311-linux_x86_64.whl
llama-index-embeddings-huggingface==0.5.2
llama-index-vector-stores-postgres==0.4.2
pdfplumber==0.10.3
minio==7.1.17
gunicorn==21.2.0
uvicorn==0.27.1
